Fix `copy_mantissa`

+++ aonfp/detail/detail.hpp
@@ -39,21 +39,21 @@ template <> inline void copy_mantissa<uint64_t>(uint64_t& mantissa, const double
 	mantissa = (*reinterpret_cast<const uint64_t*>(&v)) << 12;
 }
 template <> inline void copy_mantissa<uint32_t>(uint32_t& mantissa, const double v) {
-	const auto mantissa_bs = (*reinterpret_cast<const uint32_t*>(&v)) & 0xffffffffffffflu;
+	const auto mantissa_bs = (*reinterpret_cast<const uint64_t*>(&v)) & 0xffffffffffffflu;
 	const auto r_s = (mantissa_bs & 0x80000lu) << 1;
 	const auto mantissa_bs_a = mantissa_bs + r_s;
 	const auto digit_up = (mantissa_bs_a & 0x10000000000000lu) >> 52;
 	mantissa = mantissa_bs_a >> (20 + digit_up);
 }
 template <> inline void copy_mantissa<uint16_t>(uint16_t& mantissa, const double v) {
-	const auto mantissa_bs = (*reinterpret_cast<const uint32_t*>(&v)) & 0xffffffffffffflu;
+	const auto mantissa_bs = (*reinterpret_cast<const uint64_t*>(&v)) & 0xffffffffffffflu;
 	const auto r_s = (mantissa_bs & 0x800000000lu) << 1;
 	const auto mantissa_bs_a = mantissa_bs + r_s;
 	const auto digit_up = (mantissa_bs_a & 0x10000000000000lu) >> 52;
 	mantissa = mantissa_bs_a >> (36 + digit_up);
 }
 template <> inline void copy_mantissa<uint8_t>(uint8_t& mantissa, const double v) {
-	const auto mantissa_bs = (*reinterpret_cast<const uint32_t*>(&v)) & 0xffffffffffffflu;
+	const auto mantissa_bs = (*reinterpret_cast<const uint64_t*>(&v)) & 0xffffffffffffflu;
 	const auto r_s = (mantissa_bs & 0x80000000000lu) << 1;
 	const auto mantissa_bs_a = mantissa_bs + r_s;
 	const auto digit_up = (mantissa_bs_a & 0x10000000000000lu) >> 52;
